[doc] switch to msmb_theme, as it is up to date with readthedocs theme.
Our approach of hacking the versions combobox into the original theme was
crude and is fairly out of sync now, since it required continous rebasing
on the upstream code.

The folks at msmbuilder solved this by extending the original theme in sphinx
theme defined manner, not by forking the code.

Solves some problems like showing a scrolling bar for the menue even if there
is no content to scroll (firefox).

[ci skip]