[serialization] do not allow to load recent models to be loaded with old version
If the _serialize_version of the model is more recent than the one in the current
version of the software, we do not have a chance to interpolate correctly.

Since this would result in inconsitent models, we simply do not allow for this
and raise an informative error, which states which is the minimum version of
PyEMMA to load it again.