[sklearn] if an already estimated Estimator is passed, deep copy it
This is mandatory for parameter studies, which clones estimators to
provide a "fresh" and independent blue print.

Eg. there was a bug in HMM.cktest(), which does a parameter study over a
series of lag times and cloning led to forgetting derived attributes
like the count matrix of the initial model.

This is a ugly workaround for this situation and should be properly
fixed in the future, because deep copying the object multiplies the
memory footprint by every copy taken.