store json pickled model in hdf5, which contains all the arrays.
this seems to be the most efficient way of doing this stuff.
It also allows for easy versioning, if the same file is written multiple
times (and could potentially contain multiple models).

The arrays are being grouped in the current model group. This way one
could easily get rid of obsolete models.