[doc] filter stupid widget notice from nbconverted output.
In ipywidgets-7 there is a serialization method widget states. However,
there is no way to clear this state easily prior exporting (clear-all
or nothing, which we do not want, because we want to keep plots etc.).
So this hack now kicks out output elements, if they are part of this
generated widget output.

[ci skip]