[ML-HMSM] fix the setting of nstates (got overriden by MSM.nstates.fset)
The ML-HMSM class overrides the meaning of nstates defined in
the base class (MSM) from being a model parameter to be an input
parameter for estimation. By defining our own setter in the estimator
we avoid that this input parameter gets overriden (eg. by setting self.P
to None).

Changed its_all_estimator tests to run in parallel as it is a stronger
test, because it implies serialization and estimation.

fixes #1266