[mlmsm] fix connected sets for bayesian MSMs
These sets were not computed on the same count matrix as the one used
to compute the active set.

This had lead to differences between the largest connected set and
the shape of P and pi. When the ergodic cutoff kicked out low populated
states in during count matrix estimation, the connected sets were estimated
on a wrong measure, because it used the full count matrix instead of the
truncated one.